+++ frontend/app/chat/[chatId]/components/ChatInput/MicButton.tsx
