+++ frontend/app/chat/[chatId]/components/ChatInput/index.tsx
