+++ frontend/app/chat/[chatId]/components/ChatMessages/components/ChatMessage.tsx
@@ -35,10 +35,7 @@ export const ChatMessage = forwardRef(
           {speaker}
         </span>
         <>
-          <ReactMarkdown
-            // remarkRehypeOptions={{}}
-            className="prose dark:prose-invert ml-[6px]  mt-1"
-          >
+          <ReactMarkdown className="prose dark:prose-invert ml-[6px]  mt-1">
             {text}
           </ReactMarkdown>
         </>
