+++ frontend/app/chat/[chatId]/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./ChatInput";
+export * from "./ChatMessages";
+export * from "./ChatMessages/components/ChatMessage";
