+++ frontend/app/chat/components/index.ts
@@ -1,4 +1,1 @@
-export * from "./ChatInput";
-export * from "./ChatMessages";
-export * from "./ChatMessages/components/ChatMessage";
 export * from "./ChatsList";
