+++ frontend/lib/context/index.tsx
@@ -1,2 +1,3 @@
 export * from "./BrainProvider";
+export * from "./ChatProvider";
 export * from "./FeatureFlagProvider";
